Need a JWT to send a review + error message if unauthorized
