--- index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="/static/js/loginAuth.js" defer></script>
-    <title>test</title>
-  </head>
-  <body>
-    <input name="username" />
-    <input name="password" />
-    <button onclick="run()">Run</button>
-  </body>
-</html>
