+++ views/movie.ejs
@@ -50,6 +50,10 @@
         <!-- Namn -->
         <label class="review-form__label" for="author">Vad är ditt namn/alias:</label>
         <input class="review-form__input" type="text" id="author" name="author" required />
+
+        <!-- Lösenord -->
+         <label class="review-form__label" for="password">Vad är ditt lösenord:</label>
+         <input class="review-form__input" type="password" id="password" name="password" required />
   
         <!-- Skicka-knapp -->
         <button class="review-form__submit" type="submit">Skicka recensionen</button>
@@ -59,8 +63,7 @@
     </div>
   </dialog>
   
-
-    <script src="/static/js/reviews.js"></script>
+    <script src="/static/js/reviews.js" type="module"></script>
   </div>
 </dialog>
 <!-- End Modal-dialog -->
